Change Histogram bin count/sum type to uint64_t (via a typedef for future-proofing) so as to avoid integer overflows observed on Frontier for large numbers of ranks and functions called a large number of times (for 4K ranks it only takes 1M calls of a given function to overflow uint32_t!)

+++ test/unit_tests/ad/HBOSOutlier.cpp
@@ -152,7 +152,7 @@ TEST(HBOSADOutlierTestComputeOutliersWithoutPS, Works){
   std::list<ExecData_t> call_list3;  //aka CallList_t
   for(int i=0;i<N;i++){
     double val = i==N-1 ? 10000 : double(dist(gen)); //outlier on N-1
-    unsigned int start = 1000*(i+1);
+    double start = 1000*(i+1);
     call_list3.push_back( createFuncExecData_t(0,0,0,  func_id, "my_func", start,val) );
     runtimes.push_back(val);
     if(i==N-1){
